implemented switch case for token type; completed some documentation
